Organize code.

+++ Framework/Common/Allocator.hpp
@@ -49,17 +49,17 @@ namespace Panda{
 		Allocator& operator=(const Allocator& rhs);
 		
 	private:
-		uint32_t		m_BlockSize;			// 块的尺寸
-		uint32_t		m_PageSize;				// 页的尺寸
-		uint32_t		m_BlockCountPerPage;	// 每页的块数
-		uint32_t		m_AlignmentSize;		// 对齐尺寸
+		size_t		m_BlockSize;			// 块的尺寸
+		size_t		m_PageSize;				// 页的尺寸
+		size_t		m_BlockCountPerPage;	// 每页的块数
+		size_t		m_AlignmentSize;		// 对齐尺寸
 		
 		
 		BlockHeader*	m_pFreeBlockList;		// 空余块列表
-		uint32_t		m_FreeBlockCount;		// 空余的块数
-		uint32_t		m_BlockCount;			// 总块数
+		size_t		m_FreeBlockCount;		// 空余的块数
+		size_t		m_BlockCount;			// 总块数
 		
 		PageHeader*		m_pPageList;			// 页列表
-		uint32_t		m_PageCount;			// 页数
+		size_t		m_PageCount;			// 页数
 	};
 }
